Initial Upload
Go on a date with yourself

The mirror rules: You read this text as if the spirits in your mirror had written it for you – and then twist it as if you were one of them.

The mirror idea: To develop a better understanding of your creativity.

The mirror's goal: To thoroughly dust the shelves in your mind and thereby polish your screen.

The mirror bonus: Recognizing yourself in the person opposite you.

The Mirror Special: The images in your mirror recognize themselves in you.


Are you already in? Whether you like it or not! Whether werewolf, vampire or zombie – mirrorland.ch is made by you and for you.

At mirrorland.ch we bring real people like you to life and lure the ghosts out from behind the dusty shelves of your thoughts.

We directly target your reason and open the portal to your mind in order to sneak into your world, not through cold logic but through your hard-hitting imagination.

If that seems too arrogant to you and you don't know whether to laugh or cry, – read the message you naively wrote yourself and invite yourself to your own funeral.